+++ src/test/java/timelessodyssey/states/MenuStateTest.java
@@ -37,7 +37,7 @@ public class MenuStateTest {
 
     @Test
     public void MainMenuStep() throws IOException, URISyntaxException, FontFormatException {
-        long time = 0;
+        double time = 0;
         Mockito.when(gui.getNextAction()).thenReturn(GUI.Action.NONE);
         this.state = new MainMenuState(model){
             @Override
